# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Palestine Science Olympiad Portal API",
        "description": "RESTful API for managing science olympiad competitions, registrations, and users",
        "contact": {
            "name": "PSOP Support",
            "email": "info@psop.ps"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:82/psop/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://psop.ps/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/competitions": {
            "get": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Get all competitions",
                "description": "Returns list of all active competitions",
                "operationId": "f6fe99e5239c83221bef1f4fef724c8c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mathematics",
                                "informatics",
                                "physics",
                                "chemistry",
                                "biology",
                                "ai",
                                "cybersecurity"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name_ar": {
                                                        "type": "string"
                                                    },
                                                    "name_en": {
                                                        "type": "string"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "category": {
                                                        "type": "string"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Create new competition",
                "description": "Create a new competition (Admin only)",
                "operationId": "b6511b62ec1fd3bcf8b0c91ad1c64679",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_ar",
                                    "name_en",
                                    "code",
                                    "category"
                                ],
                                "properties": {
                                    "name_ar": {
                                        "type": "string",
                                        "example": "الأولمبياد الدولي للرياضيات"
                                    },
                                    "name_en": {
                                        "type": "string",
                                        "example": "International Mathematical Olympiad"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "IMO"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "mathematics"
                                    },
                                    "description_ar": {
                                        "type": "string"
                                    },
                                    "description_en": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Competition created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/competitions/{id}": {
            "get": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Get competition by ID",
                "description": "Returns a single competition",
                "operationId": "8532a4e433f4aa7e55128312fbea8699",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Competition ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Competition not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Update competition",
                "description": "Update an existing competition (Admin only)",
                "operationId": "1848680af8af8aebcf98960740a1a3bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Competition ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_ar": {
                                        "type": "string"
                                    },
                                    "name_en": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Competition updated successfully"
                    },
                    "404": {
                        "description": "Competition not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Competitions"
                ],
                "summary": "Delete competition",
                "description": "Delete a competition (Admin only)",
                "operationId": "a6dd554ffe7d4e285f3cb62eec430dd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Competition ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Competition deleted successfully"
                    },
                    "404": {
                        "description": "Competition not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}